fix(bbox): resolve prospector findings
- Break the logger.models -> bbox_tools cyclic import by moving
  invalidate_bbox_cache into instance.py (already in logger.models); bbox_tools
  now only holds compute_instance_bbox and is imported solely by the viewsets.
- Add the missing blank line after get_bbox_cache_ttl (pycodestyle E305).
