+++ onadata/libs/utils/bbox_tools.py
@@ -11,12 +11,6 @@ JSON (same semantics `form_tiles()` applies server-side).
 from django.contrib.gis.db.models import Extent
 
 from onadata.apps.logger.models import Instance
-from onadata.libs.utils.cache_tools import (
-    DATAVIEW_BBOX_CACHE,
-    MERGED_XFORM_BBOX_CACHE,
-    XFORM_BBOX_CACHE,
-    safe_cache_delete,
-)
 from onadata.libs.utils.dataview_filters import apply_filters
 
 
@@ -49,28 +43,3 @@ def compute_instance_bbox(xform_ids, dataview=None):
     if not extent:
         return None
     return list(extent)
-
-
-def invalidate_bbox_cache(xform_id):
-    """Bust cached bbox responses affected by a change to ``xform_id``.
-
-    Clears the form's own bbox plus every DataView and MergedXForm whose extent
-    includes this xform, so the next request recomputes from current rows. Run
-    from the Instance signals on submission create, edit, and delete.
-    """
-    # pylint: disable=import-outside-toplevel
-    from onadata.apps.logger.models import DataView, MergedXForm
-
-    safe_cache_delete(f"{XFORM_BBOX_CACHE}{xform_id}")
-
-    dataview_ids = DataView.objects.filter(xform_id=xform_id).values_list(
-        "pk", flat=True
-    )
-    for dataview_id in dataview_ids:
-        safe_cache_delete(f"{DATAVIEW_BBOX_CACHE}{dataview_id}")
-
-    merged_ids = MergedXForm.objects.filter(xforms__pk=xform_id).values_list(
-        "pk", flat=True
-    )
-    for merged_id in merged_ids:
-        safe_cache_delete(f"{MERGED_XFORM_BBOX_CACHE}{merged_id}")
